Add files via upload
Welcome to NEO's file search utility. I am still working on this little side project but figured its basically working now so I may as well share it. 

This script (in its current state) will search the contents of all files contained in the target folder and all subfolders for words contained in keywords.txt. Files that match will output the results to the working directory in a file called NEOsearchKEYWORDS.txt

Version: 0.1
Date: 17/06/2020
Functionality soon to be added:
- Optional subfolder searching
- Evaluate directory and file permissions
- Disable keyword search
- Auto crack PW hashes upon detection